+++ src/misc/optionsGatherer/types.ts
@@ -1,2 +1,3 @@
 export * from './codeOptionsGatherer';
 export * from './dynamicOptionsGatherer';
+export * from './excludingOptionsGatherer';